+++ RowDaBoat-x64barebones-d4e1c147f975/Kernel/include/functions.h
@@ -4,9 +4,9 @@
 #include <stdint.h>
 #include <lib.h>
 
-void hex_to_str(char *char_num, int num);
+void hex_to_str(char *char_num, uint64_t num);
 void dec_to_str(char *char_num, int num);
-int hex_num_length(int num);
+int hex_num_length(uint64_t num);
 int dec_num_length(int num);
 uint32_t strlen(char *str);
 char getBit(uint16_t value, int n);
